+++ example/lib/common/data/tu_chong_source.dart
@@ -1,5 +1,6 @@
 import 'dart:convert' show json;
 import 'dart:math';
+import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 
@@ -10,28 +11,6 @@ void tryCatch(Function f) {
     debugPrint('$e');
     debugPrint('$stack');
   }
-}
-
-T asT<T>(dynamic value) {
-  if (value is T) {
-    return value;
-  }
-  if (value != null) {
-    final String valueS = value.toString();
-    if (0 is T) {
-      return int.tryParse(valueS) as T;
-    } else if (0.0 is T) {
-      return double.tryParse(valueS) as T;
-    } else if ('' is T) {
-      return valueS as T;
-    } else if (false is T) {
-      if (valueS == '0' || valueS == '1') {
-        return (valueS == '1') as T;
-      }
-      return bool.fromEnvironment(value.toString()) as T;
-    }
-  }
-  return null;
 }
 
 class TuChongSource {