+++ example/lib/pages/main_page.dart
@@ -1,8 +1,9 @@
 import 'package:example/example_routes.dart';
 import 'package:extended_image/extended_image.dart';
+import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:ff_annotation_route/ff_annotation_route.dart';
+import 'package:ff_annotation_route_core/ff_annotation_route_core.dart';
 import 'package:oktoast/oktoast.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:collection/collection.dart';
@@ -24,9 +25,9 @@ class MainPage extends StatelessWidget {
 
     routesGroup.addAll(groupBy<DemoRouteResult, String>(
         routeNames
-            .map<RouteResult>((String name) => getRouteResult(name: name))
-            .where((RouteResult element) => element.exts != null)
-            .map<DemoRouteResult>((RouteResult e) => DemoRouteResult(e))
+            .map<FFRouteSettings>((String name) => getRouteSettings(name: name))
+            .where((FFRouteSettings element) => element.exts != null)
+            .map<DemoRouteResult>((FFRouteSettings e) => DemoRouteResult(e))
             .toList()
               ..sort((DemoRouteResult a, DemoRouteResult b) =>
                   b.group.compareTo(a.group)),
@@ -190,5 +191,5 @@ class DemoRouteResult {
 
   final int order;
   final String group;
-  final RouteResult routeResult;
+  final FFRouteSettings routeResult;
 }