+++ lib/src/editor/extended_image_crop_layer.dart
@@ -28,14 +28,12 @@ class ExtendedImageCropLayer extends StatefulWidget {
     this.editorConfig,
     Key key,
     this.fit,
-    this.cornerPainter,
   }) : super(key: key);
 
   final EditActionDetails editActionDetails;
   final EditorConfig editorConfig;
   final Rect layoutRect;
   final BoxFit fit;
-  final ExtendedImageCropLayerCornerPainter cornerPainter;
   @override
   ExtendedImageCropLayerState createState() => ExtendedImageCropLayerState();
 }
@@ -92,21 +90,19 @@ class ExtendedImageCropLayerState extends State<ExtendedImageCropLayer>
     }
     final EditorConfig editConfig = widget.editorConfig;
     ExtendedImageCropLayerCornerPainter cornerPainter;
+    final Color primaryColor = Theme.of(context).primaryColor;
     if (widget.editorConfig.cornerPainter == null) {
       cornerPainter = ExtendedImageCropLayerPainterNinetyDegreesCorner(
-        color: Theme.of(context).primaryColor,
+        color: primaryColor,
       );
     } else {
       cornerPainter = widget.editorConfig.cornerPainter.cornerColor == null
-          ? widget.editorConfig.cornerPainter
-              .copyWith(color: Theme.of(context).primaryColor)
+          ? widget.editorConfig.cornerPainter.copyWith(color: primaryColor)
           : widget.editorConfig.cornerPainter;
     }
-    // TODO(radomir9720): replace "(widget.editorConfig.cornerColor ?? widget.cornerPainter.cornerColor)" with "widget.cornerPainter.cornerColor". So as not to break backward compatibility, temporarily will save [widget.editorConfig.cornerColor] property.
-    // Property [cornerColor] after v1.1.2 was marked as deprecated.
-    final Color cornerColor = (widget.editorConfig.cornerColor ??
-            widget.cornerPainter?.cornerColor) ??
-        Theme.of(context).primaryColor;
+
+    final Color cornerColor = cornerPainter.cornerColor;
+
     final Color maskColor = widget.editorConfig.editorMaskColorHandler
             ?.call(context, _pointerDown) ??
         defaultEditorMaskColorHandler(context, _pointerDown);
@@ -116,12 +112,7 @@ class ExtendedImageCropLayerState extends State<ExtendedImageCropLayer>
       painter: ExtendedImageCropLayerPainter(
           cropRect: cropRect,
           cornerPainter: cornerPainter,
-          // TODO(radomir9720): remove [cornerColor] property. After v1.1.2 was marked as deprecated.
-          // ignore: deprecated_member_use_from_same_package
           cornerColor: cornerColor,
-          // TODO(radomir9720): remove [cornerSize] property. After v1.1.2 was marked as deprecated.
-          // So as not to break backward compatibility, temporarily will save [widget.editorConfig.cornerSize] property.
-          // ignore: deprecated_member_use_from_same_package
           cornerSize: editConfig.cornerSize ??
               (cornerPainter is ExtendedImageCropLayerPainterNinetyDegreesCorner
                   ? cornerPainter.cornerSize
@@ -569,15 +560,10 @@ class ExtendedImageCropLayerState extends State<ExtendedImageCropLayer>
 
 class ExtendedImageCropLayerPainter extends CustomPainter {
   ExtendedImageCropLayerPainter({
-    @required
-        this.cropRect,
+    @required this.cropRect,
     this.lineColor,
-    @Deprecated('Use cornerPainter instead. The feature was deprecated after v1.1.2.')
-        // ignore: deprecated_member_use_from_same_package
-        this.cornerColor,
-    @Deprecated('Use cornerPainter instead. The feature was deprecated after v1.1.2.')
-        // ignore: deprecated_member_use_from_same_package
-        this.cornerSize,
+    this.cornerColor,
+    this.cornerSize,
     this.lineHeight,
     this.maskColor,
     this.pointerDown,