+++ lib/src/extended_image_typedef.dart
@@ -98,3 +98,6 @@ typedef CanScaleImage = bool Function(GestureDetails details);
 
 ///Call when GestureDetails is changed
 typedef GestureDetailsIsChanged = void Function(GestureDetails details);
+
+///Call when EditActionDetails is changed
+typedef EditActionDetailsIsChanged = void Function(EditActionDetails details);